Some Changes in Group and Forever(NPM)
If not for testing purpose, please start the server as an admin with the
command "forever start ./bin/www".

You will no longer see the normal running screen of the server, but if
you want to see the child process for the server running type in
"forever list".

To stop the server, find out wich ID the child process has via "forever
list"(most likely "0") and kill it with "forever stop [ID]" ex. "forever
stop 0".